+++ plugins/src/main/java/shapes/Square.java
@@ -8,7 +8,7 @@ import javafx.scene.paint.Paint;
 public class Square {
 
     @annotations.shapes.Square
-    public void render(GraphicsContext gc, Paint color, int x, int y, int w, int h) {
+    public void render(GraphicsContext gc, Paint color, double x, double y, int w, int h) {
         Paint save = gc.getFill();
         gc.setFill(color);
         gc.fillRect(x, y, w, h);
